Remove agent/connector GUI dependency
Previously the authentication/authorization was all html/javascript driven,
with complex selenium drivers to emulate that in a bash console. Added a pkce.go
which basically authenticates using REST APIs, ie "textually". This work
is required to package the connector in a form usable by customer, we have to
give customer a simple text based option to authenticate