Re-onboard the agent/connector if gateway tunnel flaps too many times
If the tunnel to the gateway is flapping too many times, that "maybe" because
the IT admin changed some agent/connector parameters on the controller and hence the gateway is
rejecting the tunnel being onboarded with the old parameters. In this case onboard again and
get the new parameters and try a tunnel formation again.

The whole thing of re-onboarding etc.. is kept outside the core agent - the core agent just
gives stats that tunnel has flapped N times. And the agent 'management/UI' plane should look
at this stats and do a re-onboarding as per its own idea of how many flaps is too many. Right
now only the connector and the docker agent is doing this re-onboarding, eventually both
android and apple will have to do the same too