maco catalina 10.15 boringssl fails to validate certificate
The root cause is this: https://developer.apple.com/forums/thread/127779

The workaround they suggest is to enable the nextensio packet tunnel system extension to read/write
to /var/ folder - and as the link points out, apple will NOT accept that as a solution in the final
app store app. So for 10.15 the ONLY solution will be to somehow manually specify/add the certificate
in the rust native tls apis (which call into apple boringssl). For now since we are far away from an
app store app, I am just adding the read/write to the entitlements, we will need to revisit this when
we are ready to submit to app store